% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{cwlpack}
\alias{cwlpack}
\title{require sbpack to wrap cwl workflow}
\usage{
cwlpack(cwl, name = NULL)
}
\arguments{
\item{cwl}{The Rcwl object.}

\item{name}{The}
}
\description{
require sbpack to wrap cwl workflow
}
