% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{get_app}
\alias{get_app}
\title{get App details}
\usage{
get_app(project, app_id, key = NULL)
}
\arguments{
\item{project}{The project ID in the following format: {project_owner}/{project}.}

\item{app_id}{The ID for the app you are querying.}

\item{key}{The Authentication Token.}
}
\description{
get App details
}
