% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{new_task}
\alias{new_task}
\title{create new task}
\usage{
new_task(
  project,
  app,
  inputs,
  batch_input = NULL,
  batch_by = NULL,
  key = NULL,
  extlist = NULL
)
}
\arguments{
\item{project}{The project ID in the following format: {project_owner}/{project}.}

\item{app}{The ID for the app you are querying.}

\item{inputs}{The list of input parameters for the App.}

\item{batch_input}{The ID from the inputs for batch run.}

\item{batch_by}{The type of batch.}

\item{key}{The Authentication Token.}

\item{extlist}{The other requests in a list.}
}
\value{
A response list.
}
\description{
https://docs.cancergenomicscloud.org/docs/create-a-new-task
}
